chore: restructures packages to comply with Go layout conventions
