+++ Makefile
@@ -1,8 +1,8 @@
 GO := go
 
-GO_BUILD_PACKAGES := ./main/...
+GO_BUILD_PACKAGES := ./cmd/emporous-fuse
 GO_BUILD_BINDIR := ./bin
-GO_BUILD_BIN := emporous-fuse-go
+GO_BUILD_BIN := emporous-fuse
 GIT_COMMIT := $(or $(SOURCE_GIT_COMMIT),$(shell git rev-parse --short HEAD))
 GIT_TAG :="$(shell git tag | sort -V | tail -1)"
