+++ cmd/emporous-fuse/commands/log/log.go
